abstracted BuildProductCarousel out of CategoryProductCarousel
